Create Get-VMwareInformation.ps1
This is a collection of functions for reporting information about hosts and VM's.  There is also an update-vmtools that will verify the vm is powered on, update the tools without reboot and then append the timestamped result with name of vm to a log file.  These functions have been changed a bit so I still need to test them.